requirements.txt
django
django-rest-framework
requests
unidecode
django-cors-headers
pillow